Rename Raspi-Pi PhotoBooth to Raspberry Pi 3 B+ PhotoBooth
First ever project done using the Python programming language. A rather confusing and intriguing process of trial and error. It took approximately 40+ hours to get this much code in and to make the whole hardware and software function properly, but it was an amazing learning experience. I got a better feel and understanding of how 'class' functions work in Python and what it takes in order for them to all work in hand with each other to create a completed system. This is where I learned the trick to coding and that is to take everything into small bites and learn how each part works before trying to put them all together without knowing what goes where. The most entertaining part of this project was creating a Twitter bot with a selected API code in which it provided 'keys' that linked the Twitter social platform to the Raspberry Pi 3; thus, it allowed users to take photos and automatically upload it to the cloud after giving users the opportunity to caption their newly stitched photo.